+++ src/apps/http/mod.rs
@@ -1,3 +1,0 @@
-mod caddy;
-
-pub use caddy::Caddy;